fix(ring/cad): atomic iWrite/iRead ring indices (C1), atomic CAD flags (B3)
C1: iWrite/iRead volatile int -> std::atomic<uint8_t> (consistent with loraWrite).
addRingPointer(volatile int&) call for the tx ring inlined since atomic cannot
bind to the reference; DEBUG_MSG_VAL gets an explicit (int) cast. All other
accesses use atomic's implicit (seq_cst) operators.
B3: nRF52 CAD flags volatile bool -> std::atomic<bool>; publish in OnCadDone uses
memory_order_release, reads use seq_cst (>= acquire).

+++ src/nrf52/nrf_eth.cpp
@@ -636,7 +636,7 @@ void NrfETH::fillUDP_RING_BUFFER(uint8_t buffer [UDP_TX_BUF_SIZE], uint16_t rx_b
   ringBuffer[iWrite][1] = 0xFF;
   memcpy(ringBuffer[iWrite] + 2, buffer, rx_buf_size);
 
-  DEBUG_MSG_VAL("RADIO", iWrite, "fill LORA Send:");
+  DEBUG_MSG_VAL("RADIO", (int)iWrite, "fill LORA Send:");
 
   iWrite++;
   if (iWrite >= MAX_RING) // if the buffer is full we start at index 0 -> take care of overwriting!
